rpcutils: move universe gRPC marshal functionality from taprpc
Move universe related gRPC marshalling functionality from the taprpc
package to the rpcutils package as preparation for future
modularization.
